// src/types/regCert/reg/international.tsx
export interface DataType {
  key: string;
  boatSymbol: string;
  boatType: string;
  regisDate: string;
  regisOffice: string;
  occupation: string[];
  harbourCode: string;
  inspectDate: string;
  inspectPlace: string;
  ownerInfo: string;
  ownerNID: string;
  ownerPhone: string;
  captain: string;
  captainPassport: string;
  captainPhone: string;
  crewList: string[];
  licenseNumber: string;
  fishType: string[];
  fishingGrounds: string;
  seasons: string;
  output: string;
}